backend/doccollab/requirements.txt
asgiref
async-timeout
attrs
autobahn
Automat
beautifulsoup4
certifi
cffi
channels
channels_redis
charset-normalizer
constantly
cryptography
daphne
dj-database-url
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
hyperlink
idna
incremental
msgpack
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
PyJWT
pyOpenSSL
python-dotenv
redis
requests
service-identity
soupsieve
sqlparse
Twisted
txaio
typing_extensions
tzdata
urllib3
zope.interface
